requirements.txt
celery==5.2.7
Django==3.2.4
django-import-export==3.0.2
django-rest-passwordreset==1.1.0
django-throttling==0.0.1
djangorestframework==3.14.0
pytz==2021.3
PyYAML==5.1.2
requests==2.22.0
typing-extensions==3.7.4
ujson==1.35
psycopg2-binary==2.9.5
redis==4.5.1
async-timeout==4.0.2